3880 moving PageStateTabs to the tabs folder
